refactor(word): move shared value models out of dto
