+++ src/main/java/com/linglevel/api/word/dto/WordResponse.java
@@ -1,6 +1,8 @@
 package com.linglevel.api.word.dto;
 
 import com.linglevel.api.i18n.LanguageCode;
+import com.linglevel.api.word.model.Meaning;
+import com.linglevel.api.word.model.RelatedForms;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
