+++ src/main/java/com/linglevel/api/word/model/Meaning.java
@@ -1,5 +1,6 @@
-package com.linglevel.api.word.dto;
+package com.linglevel.api.word.model;
 
+import com.linglevel.api.word.dto.PartOfSpeech;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
