+++ src/main/java/com/linglevel/api/word/model/RelatedForms.java
@@ -1,4 +1,4 @@
-package com.linglevel.api.word.dto;
+package com.linglevel.api.word.model;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
